chore: Fix docker builds
because of how Mac mounts work the volume mount scheme without a named volume
won't work. This is sacrificing the ability to hot reload into the
docker container but to me that is probably not worth the build problems
mounting the app directly can cause anyway

especially if it's being deployed in non ephemeral environments. If the
container hot reload is wanted, a docker-compose-dev.yml is an option